Extract shared session actor
