+++ crates/cleat/src/lib.rs
@@ -3,6 +3,7 @@ pub mod cast_reader;
 pub mod cli;
 pub mod da;
 pub mod duration_parser;
+mod host;
 mod http_uds;
 pub mod keys;
 pub mod platform;
